Move test setup related files to tests directory.
